misc: small refactor
