+++ sennaar-rs/src/cpl/mod.rs
@@ -1,2 +1,6 @@
 mod expr;
+mod ty;
+mod decl;
 pub use expr::*;
+pub use ty::*;
+pub use decl::*;
